Revert "Refactor contentScripts and errors.js, add tests"
This reverts commit 15f23c9c39d5fc2e8c6014deb197d0d9e34f3c47.
