--- src/js/index.js
@@ -1,4 +1,0 @@
-import {activate} from './contentScripts'
-
-
-await activate()
